"Nruns" variable in the runlist is no longer needed..
Figured out an easy way to plot histograms as a fn. of run # without knowing the # runs being analyzed a priori. Until now I was handling this issue in a brute-force way not it should be more elegant. It is no longler necessary to keep track a variable, Nruns, in the run list anymore.